Fixes persistent context menu creation error "cannot create duplicate item" in service worker (#11)
Due to the contextMenu creation code living at the top level of the
service worker, that code runs every time the service worker is invoked.

I noticed it due to my other PR causing the service worker to run every
time a new tab is opened :P its a very persistent and annoying error,
but it doesn't break anything.

Its a simple fix, this contextMenu creation code should be housed within
the installation event callback, specifically when the event is INSTALL,
so it only runs on install.